+++ src/main/java/com/seoultech/ecc/study/datamodel/ExpressionRedis.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ExpressionRedis {
-    private Long expressionId;
+    private Integer expressionId;
     private String question;
     private String english;
     private String korean;
